fix qr code sizing issue

+++ lib/ui/widgets/QrCodePanel.dart
@@ -181,7 +181,7 @@ class QrCodePanel extends StatefulWidget with AnalyticsInfo { //TBD localize
 }
 
 class _QrCodePanelState extends State<QrCodePanel> {
-  static final double _imageSize = 1024;
+  static final int _imageSize = 1024;
   Uint8List? _qrCodeBytes;
 
   @override
@@ -287,7 +287,7 @@ class _QrCodePanelState extends State<QrCodePanel> {
     String qrContent = StringUtils.ensureNotEmpty(_promotionUrl ?? widget.digitalCardQrCode);
     Uint8List? imageBytes;
     if (kIsWeb) {
-      ByteData? qrPainterImage = await QrPainter(data: qrContent, version: QrVersions.auto).toImageData(_imageSize);
+      ByteData? qrPainterImage = await QrPainter(data: qrContent, version: QrVersions.auto).toImageData(_imageSize.toDouble());
       imageBytes = qrPainterImage?.buffer.asUint8List();
     } else {
       imageBytes = await NativeCommunicator().getBarcodeImageData({
